# specs/018-invitations-production-readiness/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Invitations Microservice API
  description: |
    API for managing digital wedding invitations, guest lists, RSVPs, and analytics.

    ## Authentication
    All endpoints except public RSVP require JWT authentication via Bearer token
    or internal service headers (X-User-ID, X-Workspace-ID).

    ## Rate Limits
    - Default: 100 requests/minute
    - CSV Import: 5 requests/hour
    - Bulk Invite: 10 requests/hour
    - RSVP Submit: 10 requests/minute per IP
  version: 2.0.0
  contact:
    name: RawDrive Engineering
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /workspaces/{workspaceId}/invitations/{invitationId}/analytics:
    get:
      tags: [analytics]
      summary: Get analytics summary
      description: Comprehensive analytics including views, RSVP stats, and device breakdown. Cached for 10 minutes.
      operationId: getAnalyticsSummary
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/invitationId'
      responses:
        '200':
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /workspaces/{workspaceId}/invitations/{invitationId}/analytics/views:
    get:
      tags: [analytics]
      summary: Get view statistics
      operationId: getViewStats
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/invitationId'
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month, all]
            default: week
      responses:
        '200':
          description: View statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViewStats'
  /workspaces/{workspaceId}/invitations/{invitationId}/analytics/rsvp:
    get:
      tags: [analytics]
      summary: Get RSVP statistics
      operationId: getRsvpStats
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/workspaceId'
        - $ref: '#/components/parameters/invitationId'
      responses:
        '200':
          description: RSVP statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RSVPStats'

  # ============ Bulk Operations ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    workspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Workspace identifier
    invitationId:
      name: invitationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Invitation identifier
  responses:
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Access denied to this workspace
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    AnalyticsSummary:
      type: object
      properties:
        invitation_id:
          type: string
          format: uuid
        views:
          $ref: '#/components/schemas/ViewStats'
        devices:
          $ref: '#/components/schemas/DeviceStats'
        rsvp:
          $ref: '#/components/schemas/RSVPStats'
        generated_at:
          type: string
          format: date-time
    ViewStats:
      type: object
      properties:
        total_views:
          type: integer
        unique_visitors:
          type: integer
        views_today:
          type: integer
        views_this_week:
          type: integer
        views_this_month:
          type: integer
        peak_hour:
          type: integer
          description: Hour with most views (0-23)
        peak_day:
          type: string
          description: Day with most views
    DeviceStats:
      type: object
      properties:
        desktop:
          type: integer
        mobile:
          type: integer
        tablet:
          type: integer
        unknown:
          type: integer
        browsers:
          type: object
          additionalProperties:
            type: integer
        operating_systems:
          type: object
          additionalProperties:
            type: integer
    RSVPStats:
      type: object
      properties:
        total_guests:
          type: integer
        responded:
          type: integer
        attending:
          type: integer
        not_attending:
          type: integer
        maybe:
          type: integer
        pending:
          type: integer
        response_rate:
          type: number
          format: float
        total_plus_ones:
          type: integer
        expected_attendance:
          type: integer
          description: attending + total_plus_ones

    # ============ Bulk Invite Schemas ============
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string